pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "podforge"
version = "0.1.0"
description = "Turn any topic, URL, or document into a fully-produced podcast episode"
requires-python = ">= 3.10"
dependencies = [
    "click",
    "anthropic",
    "elevenlabs",
    "edge-tts",
    "trafilatura",
    "pdfplumber",
    "pyyaml",
    "mutagen",
    "rich",
    "httpx",
]

[project.scripts]
podforge = "podforge.cli:cli"

[dependency-groups]
dev = [
    "pytest",
    "pytest-asyncio",
]
